requirements.txt
pillow
numpy
streamlit
tensorflow
opencv-python
torch
torchvision
albumentations
matplotlib